requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
openai==1.57.0
qdrant-client==1.11.3
numpy==2.1.3
python-dotenv==1.0.1
loguru==0.7.2
slowapi==0.1.9
tenacity==8.2.3
redis==5.0.1
duckduckgo-search==5.3.0
pypdf==3.17.4
python-multipart==0.0.9

pytest==8.3.3
pytest-asyncio==0.24.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0

pandas==2.2.0
openpyxl==3.1.2

# LangGraph and LangChain dependencies
langgraph
langchain
langchain-openai
langchain-community
langchain-core

# Phase 2: Advanced Retrieval
rank-bm25==0.2.2  # Hybrid search (BM25 + embeddings)

# Phase 5: Vision & Multimodal
pillow==10.1.0  # Image processing

# Phase 7: Advanced Production Features
hiredis==2.2.3  # High-performance Redis parser